refactor(coins): move base fragment & adapter to dedicated package
