+++ app/src/main/java/com/hxl/cryptonumismatist/base/Binder.java
@@ -1,4 +1,4 @@
-package com.hxl.cryptonumismatist.utils;
+package com.hxl.cryptonumismatist.base;
 
 public interface Binder<T> {
     void bind(T item);
